move flagcx implementation to separate file
